+++ fullcontact/__init__.py
@@ -5,7 +5,7 @@ from .api.person_api import PersonApi as _PersonApi
 from .api.resolve_api import ResolveApi as _ResolveApi
 from .api.tags_api import TagsApi as _TagsApi
 from .api.audience_api import AudienceApi as _AudienceApi
-from .api.verification import EmailVerificationApi as _EmailVerificationApi
+from .api.verification_api import VerificationApi as _EmailVerificationApi
 from .config.client_config import ClientConfig as _ClientConfig
